Updated FFmpeg
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@2829 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/ffmpeg/libavcodec/dsputil.c
@@ -4142,7 +4142,7 @@ int ff_check_alignment(void){
     static int did_fail=0;
     DECLARE_ALIGNED_16(int, aligned);
 
-    if((long)&aligned & 15){
+    if((intptr_t)&aligned & 15){
         if(!did_fail){
 #if HAVE_MMX
             av_log(NULL, AV_LOG_ERROR,
